+++ packages/components/molecules/f-skeleton-loader/test-utils/component-objects/f-skeletonLoader.component.js
@@ -1,0 +1,18 @@
+const Page = require('@justeat/f-wdio-utils/src/page.object');
+const { COMPONENT } = require('./f-skeletonLoader-selectors');
+
+module.exports = class SkeletonLoader extends Page {
+    get component () { return $(COMPONENT); }
+
+    open () {
+        super.openComponent('molecule', 'skeleton-loader-component');
+    }
+
+    waitForComponent () {
+        super.waitForComponent(this.component);
+    }
+
+    isComponentDisplayed () {
+        return this.component.isDisplayed();
+    }
+};