+++ packages/components/organisms/f-checkout/src/tenants/en-GB.js
@@ -48,6 +48,13 @@ const messages = {
         }
     },
 
+    warningMessages: {
+        preOrder: {
+            title: 'Please note, this is a preorder',
+            body: 'Please check the day and time above'
+        }
+    },
+
     buttonText: 'Go to payment',
 
     userNote: {