+++ packages/components/organisms/f-registration/src/components/_tests/Registration.test.js
@@ -10,6 +10,7 @@ let wrapper;
 
 describe('Registration', () => {
     const propsData = {
+        locale: "en-GB",
         createAccountUrl: 'http://localhost/account/register',
         showLoginLink: true,
         loginUrl: '/account/register'
@@ -108,20 +109,6 @@ describe('Registration', () => {
 
             // Assert
             expect(wrapper.emitted(EventNames.VisitLoginPage).length).toBe(1);
-        });
-
-        it('should fallback to use the en-GB locale if no locale passed', () => {
-            // Arrange & Act
-            wrapper = shallowMount(Registration, {
-                propsData: {
-                    createAccountUrl: 'http://localhost/account/register',
-                    showLoginLink: true,
-                    loginUrl: '/account/register'
-                }
-            });
-
-            // Assert
-            expect(wrapper.vm.tenant).toBe('uk');
         });
     });