+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-v3.23.0
+v3.24.0
 ------------------------------
 *March 22, 2021*
 
@@ -11,6 +11,14 @@ v3.23.0
 - Refactored `jest-allure2` and `allure-reporter`
 - Added script `report:test-component:chrome` into package.json for automating component test report for organisms
 - Updated generator to reflect changes
+
+
+v3.23.0
+------------------------------
+*March 18, 2021*
+
+## Changed
+- Updated node `config.yml` to support chromedriver 89.0.0
 
 
 v3.22.0